refactor: Calculator -> CalculatorController 변경 및 메서드 분할
